Content for v6.0

+++ benchmarks/src/main/scala/co/blocke/scalajack/benchmarks/Person.scala
@@ -1,3 +1,3 @@
 package co.blocke.scalajack.benchmarks
 
-case class Person(id: Long, first_name: String, last_name: String, email: String, gender: String, ip_address: String)
+case class Person(id: Int, first_name: String, last_name: String, email: String, gender: String, ip_address: String)
